pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ad-groups-mcp"
version = "0.1.0"
description = "Read-only Active Directory group management MCP server"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "fastmcp>=2.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
